==> code/Step2_AddDataFields.ipynb <==
# Step 2: Add Data Fields

## Setup

import pandas as pd
from master_functions import *
import os

#### Load Dataset

df = pd.read_csv('https://raw.githubusercontent.com/lisavirginia/clinical-abbreviations/master/code/Step1Output.csv',
                 sep='|',
                 header=0,
                 index_col=False,
                 na_filter=False,
                 dtype=object)

df.sample(3, random_state=0)

## Add Record Identifier

#### Assign Record Identifier

assignment = 1
for index, row in df.iterrows():
    df['RecordID'].iat[index] = assignment
    assignment += 1

#### Format Record Identifier

df['RecordID'] = 'R' + (df.RecordID.map('{:06}'.format))
df.head(3)

## Add Normalized Short Forms

The normalized short form is created by:
1. converting all text to lowercase; 
2. stripping leading and trailing whitespace;
3. standardizing punctuation to an underscore.

df['NormSF'] = df['SF'].apply(normalized_short_form)

df.sample(3, random_state=0)

## Add Normalized Long Forms

Requires local installation of UMLS Lexical Variant Generation program, downloadable [here](https://www.nlm.nih.gov/research/umls/licensedcontent/umlsknowledgesources.html).

Additional documentation available [here](https://www.ncbi.nlm.nih.gov/books/NBK9680/). Installation instructions available [here](https://lexsrv2.nlm.nih.gov/LexSysGroup/Projects/lvg/current/docs/userDoc/install/install.html).

Here, we used version lvg2019 AB.

lvg_path = 'C:/Users/lvg2104/Documents/clinical-abbreviations/code/lvg2019/bin/lvg.bat'

#### Identify Non-ASCII Characters

df['ASCII'] = 'Y'
df.loc[df.LF.str.contains('[^\x00-\x7F]') == True, 'ASCII'] = 'N'

#### Extract Unique LFs

uniq_LFs = pd.Series(df.loc[df['ASCII']=='Y']['LF'].unique())

uniq_LFs.to_csv('uniq_LFs.temp',
                index=False,
                header=False,
                encoding='ascii')

#### Call Lexical Variant Generation

lvg_output = lvg(input_file='uniq_LFs.temp',
                 flow='q0:g:rs:o:t:l:B:Ct:q7:q8',
                 output_file='norm_LFs.temp',
                 lvg_path=lvg_path)

#### Extract Normalized LFs

norm_LFs = pd.read_csv('norm_LFs.temp',
                       sep='|',
                       header = None,
                       usecols = [1])

#### Match Unique & Normalized LFs

normed_df = pd.DataFrame()
normed_df['LF'] = uniq_LFs
normed_df['NormLF_temp'] = norm_LFs

#### Populate Database

df = pd.merge(df, normed_df, how='left', on='LF')
df['NormLF'] = df['NormLF_temp']

df = df.drop(columns=['NormLF_temp', 'ASCII'])

df.sample(3, random_state=0)

#### Clean Environment

df = df.replace('-No Output-', '')
df.sample(3, random_state=0)

os.remove('uniq_LFs.temp')
os.remove('norm_LFs.temp')

## Add Short Form Unique Identifier

#### Sort by SF

df = df.sort_values(by=['SF'])
df = df.reset_index(drop=True)

#### Assign SFUI

assignment = 1
for index, row in df.iterrows():
    if index == 0:
        df['SFUI'].iat[index] = assignment
    elif df['SF'].at[index] == df['SF'].at[index-1]:
        df['SFUI'].iat[index] = assignment
    else:
        assignment += 1
        df['SFUI'].iat[index] = assignment

#### Format SFUI

df['SFUI'] = 'S' + (df.SFUI.map('{:06}'.format))
df.head(5)

## Add Long Form Unique Identifier

#### Sort by LF

df = df.sort_values(by=['LF'])
df = df.reset_index(drop=True)

#### Assign LFUI

assignment = 1
for index, row in df.iterrows():
    if index == 0:
        df['LFUI'].iat[index] = assignment
    elif df['LF'].at[index] == df['LF'].at[index-1]:
        df['LFUI'].iat[index] = assignment
    else:
        assignment += 1
        df['LFUI'].iat[index] = assignment

#### Format LFUI

df['LFUI'] = 'L' + (df.LFUI.map('{:06}'.format))
df.head(5)

## Export

df = df.sort_values(by=['RecordID'])
df = df.reset_index(drop=True)

df.to_csv('Step2Output.csv',
          index=False,
          header=True,
          sep='|')
